fix: Update tests
