--- test/resources/incompatibility/ignores-nullable-for-schemas-in-31.json
@@ -1,22 +1,0 @@
-{
-  "openapi": "3.1.0",
-  "paths": {
-    "/example": {
-      "post": {
-        "responses": {
-          "200": {
-            "description": "OK",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "nullable": true,
-                  "type": "string"
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
