--- test/resources/incompatibility/type-null-in-30.json
@@ -1,21 +1,0 @@
-{
-  "openapi": "3.0.0",
-  "paths": {
-    "/example": {
-      "post": {
-        "responses": {
-          "200": {
-            "description": "OK",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "null"
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
