--- test/resources/incompatibility/types-in-array-are-unique-31.json
@@ -1,24 +1,0 @@
-{
-  "openapi": "3.1.0",
-  "paths": {
-    "/example": {
-      "post": {
-        "responses": {
-          "200": {
-            "description": "OK",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": [
-                    "string",
-                    "string"
-                  ]
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
